Pull printf and I/O helpers out into printf.c
