+++ kernel/printf.h
@@ -1,0 +1,8 @@
+#ifndef PRINTF_H
+#define PRINTF_H
+
+char getchar();
+void putchar( char c );
+void printf( const char *s, ... );
+
+#endif
